add compnent
