+++ src/components/Player.jsx
@@ -1,3 +1,5 @@
+import TrackPlayRender from "./PlayerTrackPlay";
+
 function PlayerRender() {
   return (
     <div className="bar">
@@ -33,38 +35,8 @@ function PlayerRender() {
               </div>
             </div>
 
-            <div className="player__track-play track-play">
-              <div className="track-play__contain">
-                <div className="track-play__image">
-                  <svg className="track-play__svg" alt="music">
-                    <use xlinkHref="./img/icon/sprite.svg#icon-note"></use>
-                  </svg>
-                </div>
-                <div className="track-play__author">
-                  <a className="track-play__author-link" href="http://">
-                    Ты та...
-                  </a>
-                </div>
-                <div className="track-play__album">
-                  <a className="track-play__album-link" href="http://">
-                    Баста
-                  </a>
-                </div>
-              </div>
+            <TrackPlayRender author_link="http://" author_text= "Ты та..." album_link="http://" album_text="Баста" />
 
-              <div className="track-play__like-dis">
-                <div className="track-play__like _btn-icon">
-                  <svg className="track-play__like-svg" alt="like">
-                    <use xlinkHref="./img/icon/sprite.svg#icon-like"></use>
-                  </svg>
-                </div>
-                <div className="track-play__dislike _btn-icon">
-                  <svg className="track-play__dislike-svg" alt="dislike">
-                    <use xlinkHref="./img/icon/sprite.svg#icon-dislike"></use>
-                  </svg>
-                </div>
-              </div>
-            </div>
           </div>
           <div className="bar__volume-block volume">
             <div className="volume__content">
